Use connman_dbus 0.2.x backend and update tooltip
The only visible change should be that the tooltip now shows the network
status (e.g. "online") and the WiFi signal (when applicable) in parentheses.

Internally, use the connman_dbus 0.2.x as backend to profit of the improvements
given by the dependency on lgi/GLib instead of ldbus.

The code has also been greatly simplified.